My final word on ACCC complaint
THE hard reality that are there in plainsite on the @WildDuckPhD.

Notice how even telling the cold hard facts my prose turns back
to rhyme if I am given the smallest change. I really can't explain
this state I am in.

But really folks. Wake up, smell the coffee. Read the arguments.

The facts are there staring at you.

I will believe the facts are wrong Telstra is able to publish a rebuttal
of my claims and that everyone believes is sound. Ask for log files. Ask for proof.
Ask why Telstra can't get it the answer is there waiting to be found.

I understand why my caring friends were worried and couldn't keep up
with the state I was in, I was so worked up

But really, in the cold light of day can everyone else read my
Linked In profile and not see a kernel truth.

Maybe the problem is that people like me and also my friends never
get done and so never need amends.

It is the common man who suffers this outrageout abuse. Ask them wide
and deep questions about their SMS TXT use. Ask Nate Silver - the genius to help
prepare and collect statistics on their losses.

If at the end of all this you think I am still mad, then give me a pad.
A room with no view, and where I can sit still and stew

Signed-off-by: Jon Seymour <jon@wildducktheories.com>